Fixing ThreadManager deconstructor problem
CThreadManager is not cleaning up properly.  Access violation in vector::orphan_all_unlocked_v3().  Apparently something to do with manipulation of the vector contents, ostensibly in the process() function.  Can't seem to find anything relevant.  Might be better idea to just use a flag on the CThreadJob level to indicate what status it's in rather than hopping the pointer between three different queues.